+++ packages/vroagn/src/index.iife.ts
@@ -2,6 +2,8 @@ import { iife } from '../../../.scripts/iife.js'
 
 import { create } from './library/request.js'
 
+import { cacheFetch } from './library/cache.js'
+
 import { csvParser } from './library/parsers/csv.js'
 import { iniParser } from './library/parsers/ini.js'
 import { tomlParser } from './library/parsers/toml.js'
@@ -11,6 +13,8 @@ iife([
   'vroagn',
 ], {
   create,
+
+  cacheFetch,
 
   csvParser,
   iniParser,