+++ packages/vroagn/src/index.ts
@@ -1,5 +1,7 @@
 export * from './library/request.js'
 
+export * from './library/cache.js'
+
 export * from './library/parsers/csv.js'
 export * from './library/parsers/ini.js'
 export * from './library/parsers/toml.js'